Use jlong instead of int as pointer holder
git-svn-id: svn://svn.rforge.net/org/trunk@1824 955e8658-2bef-402f-a597-8b5c18d8b76d

+++ rosuda/JRI/src/jri.h
@@ -15,8 +15,9 @@
 extern "C" {
 #endif
     
-#define SEXP2L(s) ((long)(s))
-#define L2SEXP(s) ((SEXP)((int)(s)))
+  /* jlong can always hold a pointer */
+#define SEXP2L(s) ((jlong)(s))
+#define L2SEXP(s) ((SEXP)((jlong)(s)))
         
 jstring jri_callToString(JNIEnv *env, jobject o);
